# api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    },
    {
      "url": "http://172.30.80.11:31022/",
      "description": "Server url"
    }
  ],
...
  "paths": {
    "/api/tarification/{id}/valider": {
      "put": {
        "tags": [
          "tarification-controller"
        ],
        "summary": "Valider une tafification par sont ID",
        "operationId": "validerTarification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TarificationResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/tarification/{id}/paiement": {
      "post": {
        "tags": [
          "tarification-controller"
        ],
        "summary": "Effectue le paiement d'une tarification validée",
        "operationId": "payerTarification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "La tarification n'est pas validée ou a expiré",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "409": {
            "description": "La tarification est déjà payée",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Tarification non trouvée",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "Paiement effectué avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TarificationResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/tarification/demandes/calculer-tarif": {
      "post": {
        "tags": [
          "Tarification des Demandes"
        ],
        "summary": "Calcule et enregistre un tarif pour une demande avec itinéraire",
        "operationId": "calculerTarif",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TarificationRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Données de requête invalides",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TarificationResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TarificationResponseDTO"
                }
              }
            }
          },
          "201": {
            "description": "Tarification calculé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TarificationResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/tarification/calculer": {
      "post": {
        "tags": [
          "tarification-controller"
        ],
        "summary": "Calculer une tarification à partie de demandeID et itineraireId",
        "operationId": "calculerTarif_1",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TarificationRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TarificationResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tarification/{id}": {
      "get": {
        "tags": [
          "tarification-controller"
        ],
        "summary": "Récupérer une tarification par son ID",
        "operationId": "getTarification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TarificationResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tarification/demandes/tarification/{id}": {
      "get": {
        "tags": [
          "Tarification des Demandes"
        ],
        "summary": "Récupère une tarification par son ID",
        "operationId": "getTarificationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TarificationResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "TarificationResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "demandeId": {
            "type": "integer",
            "format": "int64"
          },
          "itineraireId": {
            "type": "string"
          },
          "clientId": {
            "type": "integer",
            "format": "int64"
          },
          "chauffeurId": {
            "type": "string"
          },
          "volume": {
            "type": "number"
          },
          "poids": {
            "type": "number"
          },
          "distanceKm": {
            "type": "number",
            "format": "double"
          },
          "natureMarchandise": {
            "type": "string"
          },
          "fragile": {
            "type": "boolean"
          },
          "temperatureRequise": {
            "type": "string"
          },
          "tarifClient": {
            "type": "number"
          },
          "tarifPrestataire": {
            "type": "number"
          },
          "margeService": {
            "type": "number"
          },
          "statut": {
            "type": "string",
            "enum": [
              "EN_ATTENTE",
              "VALIDE",
              "REJETE",
              "ANNULEE",
              "EXPIRE",
              "PAYE"
            ]
          },
          "typeRoute": {
            "type": "string",
            "enum": [
              "ROUTE_NATIONALE"
            ]
          },
          "inclureRetour": {
            "type": "boolean"
          },
          "dateCreation": {
            "type": "string",
            "format": "date-time"
          },
          "dateExpiration": {
            "type": "string",
            "format": "date-time"
          },
          "dateValidation": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TarificationRequestDTO": {
        "required": [
          "demandeId",
          "itineraireId"
        ],
        "type": "object",
        "properties": {
          "demandeId": {
            "type": "integer",
            "format": "int64"
          },
          "itineraireId": {
            "type": "string"
          },
          "typeRoute": {
            "type": "string"
          },
          "inclureRetour": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}